+++ src/randomStr/__tests__/index.test.ts
@@ -1,0 +1,8 @@
+import randomStr from '../index';
+
+describe('randomStr', () => {
+  test('生成随机字符串', () => {
+    const len = 6;
+    expect(randomStr(len).length).toBe(len);
+  });
+});
